# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Web Calculator Gateway Service API",
        "title": "Web Calculator API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/calculate": {
            "post": {
                "description": "Evaluates a mathematical expression and returns the result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orchestrator"
                ],
                "summary": "Calculate mathematical expression",
                "parameters": [
                    {
                        "description": "Expression to calculate",
                        "name": "expression",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.CalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemas.CalculateResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/schemas.CannotParseExpression"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.InternalServerError"
                        }
                    }
                }
            }
        },
        "/expressions": {
            "get": {
                "description": "Returns a list of all calculated expressions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orchestrator"
                ],
                "summary": "Get all expressions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ExpressionsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.InternalServerError"
                        }
                    }
                }
            }
        },
        "/expressions/{id}": {
            "get": {
                "description": "Returns a specific expression by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orchestrator"
                ],
                "summary": "Get expression by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Expression ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ExpressionByIdResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ExpressionNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.InternalServerError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "schemas.CalculateRequest": {
            "type": "object",
            "properties": {
                "expression": {
                    "type": "string",
                    "example": "40+2"
                }
            }
        },
        "schemas.CalculateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "schemas.CannotParseExpression": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "cannot parse expression"
                }
            }
        },
        "schemas.Expression": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "result": {
                    "type": "number",
                    "example": 42
                },
                "status": {
                    "type": "string",
                    "example": "done"
                }
            }
        },
        "schemas.ExpressionByIdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "result": {
                    "type": "number",
                    "example": 42
                },
                "status": {
                    "type": "string",
                    "example": "done"
                }
            }
        },
        "schemas.ExpressionNotFound": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "expression not found"
                }
            }
        },
        "schemas.ExpressionsResponse": {
            "type": "object",
            "properties": {
                "expressions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.Expression"
                    }
                }
            }
        },
        "schemas.InternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        }
    }
}